fix:week1
